--- a/deploy/cpp_infer/readme_en.md
+++ b/deploy/cpp_infer/readme_en.md
@@ -101,12 +101,12 @@ There are two ways to obtain the Paddle prediction library, which will be descri
 
 #### 1.2.1 Download and install directly
 
-* [Paddle prediction library official website](https://paddleinference.paddlepaddle.org.cn/v2.1/user_guides/download_lib.html) provides different cuda versions of Linux prediction libraries, you can Check and **select the appropriate prediction library version** on the official website (it is recommended to select the prediction library with paddle version>=2.0.1).
+* [Paddle prediction library official website](https://paddleinference.paddlepaddle.org.cn/v2.2/user_guides/download_lib.html) provides different cuda versions of Linux prediction libraries, you can Check and **select the appropriate prediction library version** on the official website (it is recommended to select the prediction library with paddle version>=2.0.1).
 
 * Download and get a `paddle_inference.tgz` compressed package, and then unzip it into a folder, the command is as follows (taking the machine environment as gcc8.2 as an example):
 
     ```bash
-    wget https://paddle-inference-lib.bj.bcebos.com/2.1.1-gpu-cuda10.1-cudnn7-mkl-gcc8.2/paddle_inference.tgz
+    wget https://paddle-inference-lib.bj.bcebos.com/2.2.2/cxx_c/Linux/GPU/x86-64_gcc8.2_avx_mkl_cuda10.1_cudnn7.6.5_trt6.0.1.5/paddle_inference.tgz
     tar -xf paddle_inference.tgz
     ```
 
@@ -118,7 +118,7 @@ There are two ways to obtain the Paddle prediction library, which will be descri
 
     ```shell
     git clone https://github.com/PaddlePaddle/Paddle.git
-    git checkout release/2.1
+    git checkout release/2.2
     ```
 
 * After entering the Paddle directory, the compilation method is as follows.
@@ -217,7 +217,16 @@ Operation mode:
 
 Among them, `mode` is a required parameter, which means the selected function, and the value range is ['rec'], which means **video recognition** (more functions will be added in succession).
 
+Note: Note: If you want to enable the TensorRT optimization option during prediction, you need to run the following command to set the relevant path of TensorRT.
+```bash
+export PATH=$PATH:/path/to/cuda/bin
+export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/path/to/cuda/bin
+export LIBRARY_PATH=$LIBRARY_PATH:/path/to/cuda/bin
+export LD_LIBRARY_PATH=/path/to/TensorRT-x.x.x.x/lib:$LD_LIBRARY_PATH
+```
+
 ##### 1. Call video recognition:
+
 ```bash
 # run PP-TSM inference
 ./build/ppvideo rec \
@@ -291,4 +300,4 @@ I1125 08:10:45.834602 13955 autolog.h:67] preprocess_time(ms): 10.6524, inferenc
 
 ### 3 Attention
 
-* When using the Paddle prediction library, it is recommended to use the prediction library of version 2.1.0.
+* When using the Paddle prediction library, it is recommended to use the prediction library of version 2.2.2.